+++ lib/src/value.rs
@@ -16,7 +16,7 @@ use nom::{
 #[derive(Debug, PartialEq, Clone)]
 /// Value in HCL
 pub enum Value<'a> {
-    Integer(i64),
+    Integer(i128),
     Float(f64),
     Boolean(bool),
     String(String),
@@ -116,7 +116,7 @@ impl<'a> Value<'a> {
         self.len() == 0
     }
 
-    pub fn integer(&self) -> Result<i64, Error> {
+    pub fn integer(&self) -> Result<i128, Error> {
         if let Value::Integer(i) = self {
             Ok(*i)
         } else {
@@ -130,7 +130,7 @@ impl<'a> Value<'a> {
 
     /// # Panics
     /// Panics if the variant is not an integer
-    pub fn unwrap_integer(&self) -> i64 {
+    pub fn unwrap_integer(&self) -> i128 {
         self.integer().unwrap()
     }
 
@@ -489,7 +489,7 @@ where
     }
 }
 
-impl_from_value!(Integer, i64);
+impl_from_value!(Integer, i128);
 impl_from_value!(Float, f64);
 impl_from_value!(Boolean, bool);
 impl_from_value!(String, String);
